test_suite/requirements.txt
protobuf
requests
python-snappy
